feat: jwt tokenization and protected methods implemented

+++ book-service/internal/database/repository/repository.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"common/models"
+	"common/network/requests"
 	"context"
 
 	"github.com/jackc/pgx/v5"
@@ -43,19 +44,19 @@ func (repo *BookRepository) GetBook(ctx context.Context, id int64) (*models.Book
 	return &book, nil
 }
 
-func (repo *BookRepository) CreateBook(ctx context.Context, book *models.Book) (int64, error) {
-	err := repo.pool.QueryRow(ctx, "INSERT INTO books (title, description, author_id) VALUES ($1, $2, $3) RETURNING id", book.Title, book.Description, book.AuthorID).
-		Scan(&book.ID)
+func (repo *BookRepository) CreateBook(ctx context.Context, req *requests.BookCreateRequest) (int64, error) {
+	err := repo.pool.QueryRow(ctx, "INSERT INTO books (title, description, author_id) VALUES ($1, $2, $3) RETURNING id", req.Title, req.Description, req.AuthorID).
+		Scan(&req.ID)
 
 	if err != nil {
-		return book.ID, err
+		return req.ID, err
 	}
 
-	return book.ID, nil
+	return req.ID, nil
 }
 
-func (repo *BookRepository) UpdateBook(ctx context.Context, id int64, book *models.Book) (int64, error) {
-	tag, err := repo.pool.Exec(ctx, "UPDATE books SET title=$1, description=$2, author_id=$3 WHERE id=$4", book.Title, book.Description, book.AuthorID, id)
+func (repo *BookRepository) UpdateBook(ctx context.Context, id int64, req *requests.BookUpdateRequest) (int64, error) {
+	tag, err := repo.pool.Exec(ctx, "UPDATE books SET title=$1, description=$2 WHERE id=$3 AND author_id=$4", req.Title, req.Description, id, req.AuthorID)
 
 	if err != nil {
 		return 0, err
@@ -64,8 +65,8 @@ func (repo *BookRepository) UpdateBook(ctx context.Context, id int64, book *mode
 	return tag.RowsAffected(), nil
 }
 
-func (repo *BookRepository) DeleteBook(ctx context.Context, id int64) (int64, error) {
-	tag, err := repo.pool.Exec(ctx, "DELETE FROM books WHERE id=$1", id)
+func (repo *BookRepository) DeleteBook(ctx context.Context, id int64, authorId int64) (int64, error) {
+	tag, err := repo.pool.Exec(ctx, "DELETE FROM books WHERE id=$1 AND author_id=$2", id, authorId)
 
 	if err != nil {
 		return 0, err
